# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "工程认证智能体 API",
    "summary": "实验教学证据治理与持续改进平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recognition/candidates": {
      "get": {
        "tags": [
          "recognition"
        ],
        "summary": "获取识别候选列表",
        "operationId": "list_candidates_api_v1_recognition_candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按课程筛选",
              "title": "Course"
            },
            "description": "按课程筛选"
          },
          {
            "name": "risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按风险等级筛选",
              "title": "Risk"
            },
            "description": "按风险等级筛选"
          },
          {
            "name": "candidate_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按候选类型筛选",
              "title": "Candidate Type"
            },
            "description": "按候选类型筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecognitionCandidateResponse"
                  },
                  "title": "Response List Candidates Api V1 Recognition Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recognition/candidates/{candidate_id}/review": {
      "post": {
        "tags": [
          "recognition"
        ],
        "summary": "审核候选关系",
        "operationId": "review_candidate_api_v1_recognition_candidates__candidate_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionCandidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Get ability graph",
        "operationId": "get_graph_api_v1_graph_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityGraphResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/graph/edges/{edge_id}/review": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Review graph edge",
        "operationId": "review_edge_api_v1_graph_edges__edge_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "edge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Edge Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphEdgeReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityGraphEdgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/graph": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "summary": "Run graph diagnostics from the approved ability graph",
        "operationId": "run_graph_diagnostics_api_v1_diagnostics_graph_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphDiagnosticReportResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/findings": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "summary": "获取诊断发现列表",
        "operationId": "list_findings_api_v1_diagnostics_findings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按课程筛选",
              "title": "Course"
            },
            "description": "按课程筛选"
          },
          {
            "name": "risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按风险等级筛选",
              "title": "Risk"
            },
            "description": "按风险等级筛选"
          },
          {
            "name": "finding_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按发现类型筛选",
              "title": "Finding Type"
            },
            "description": "按发现类型筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiagnosticFindingResponse"
                  },
                  "title": "Response List Findings Api V1 Diagnostics Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/findings/{finding_id}/decision": {
      "post": {
        "tags": [
          "diagnostics"
        ],
        "summary": "处置诊断发现",
        "operationId": "decide_finding_api_v1_diagnostics_findings__finding_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticFindingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbilityGraphEdgeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "sourceType": {
            "type": "string",
            "title": "Sourcetype"
          },
          "reviewStatus": {
            "type": "string",
            "title": "Reviewstatus"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "aiReasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aireasoning"
          },
          "candidateId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidateid"
          },
          "reviewedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewedby"
          },
          "reviewedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewedat"
          },
          "evidenceSummary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidencesummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "target",
          "kind",
          "sourceType",
          "reviewStatus"
        ],
        "title": "AbilityGraphEdgeResponse"
      },
      "AbilityGraphNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "properties": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "code",
          "name",
          "properties"
        ],
        "title": "AbilityGraphNodeResponse"
      },
      "AbilityGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/AbilityGraphNodeResponse"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/AbilityGraphEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "AbilityGraphResponse"
      },
...
      "CandidateEvidenceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "resourceName": {
            "type": "string",
            "title": "Resourcename"
          },
          "resourceVersion": {
            "type": "string",
            "title": "Resourceversion"
          },
          "coordinate": {
            "type": "string",
            "title": "Coordinate"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          }
        },
        "type": "object",
        "required": [
          "id",
          "resourceName",
          "resourceVersion",
          "coordinate",
          "excerpt",
          "hash"
        ],
        "title": "CandidateEvidenceResponse"
      },
      "CandidateReviewRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "sourceNode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcenode"
          },
          "targetNode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetnode"
          },
          "relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          },
          "evidenceExcerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidenceexcerpt"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "CandidateReviewRequest"
      },
      "CandidateReviewStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "rejected",
          "modified"
        ],
        "title": "CandidateReviewStatus"
      },
...
      "DiagnosticEvidenceRefResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "objectName": {
            "type": "string",
            "title": "Objectname"
          },
          "objectVersion": {
            "type": "string",
            "title": "Objectversion"
          },
          "coordinate": {
            "type": "string",
            "title": "Coordinate"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          }
        },
        "type": "object",
        "required": [
          "id",
          "objectName",
          "objectVersion",
          "coordinate",
          "excerpt",
          "hash"
        ],
        "title": "DiagnosticEvidenceRefResponse"
      },
      "DiagnosticFindingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "course": {
            "type": "string",
            "title": "Course"
          },
          "type": {
            "$ref": "#/components/schemas/DiagnosticFindingType"
          },
          "risk": {
            "$ref": "#/components/schemas/DiagnosticFindingRisk"
          },
          "sourceNode": {
            "type": "string",
            "title": "Sourcenode"
          },
          "targetNode": {
            "type": "string",
            "title": "Targetnode"
          },
          "relationLabel": {
            "type": "string",
            "title": "Relationlabel"
          },
          "graphVersion": {
            "type": "string",
            "title": "Graphversion"
          },
          "rule": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Rule"
          },
          "decisionStatus": {
            "$ref": "#/components/schemas/FindingDecisionStatus"
          },
          "impact": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Impact"
          },
          "suggestedDestination": {
            "type": "string",
            "title": "Suggesteddestination"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticEvidenceRefResponse"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "course",
          "type",
          "risk",
          "sourceNode",
          "targetNode",
          "relationLabel",
          "graphVersion",
          "rule",
          "decisionStatus",
          "impact",
          "suggestedDestination",
          "evidence"
        ],
        "title": "DiagnosticFindingResponse"
      },
      "DiagnosticFindingRisk": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "DiagnosticFindingRisk"
      },
      "DiagnosticFindingType": {
        "type": "string",
        "enum": [
          "coverage-gap",
          "material-conflict",
          "structural-risk",
          "version-impact"
        ],
        "title": "DiagnosticFindingType"
      },
...
      "FindingDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "FindingDecisionRequest"
      },
      "FindingDecisionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "dismissed",
          "converted"
        ],
        "title": "FindingDecisionStatus"
      },
...
      "GraphDiagnosticReportResponse": {
        "properties": {
          "graphVersion": {
            "type": "string",
            "title": "Graphversion"
          },
          "generatedAt": {
            "type": "string",
            "title": "Generatedat"
          },
          "overallCoverageRate": {
            "type": "number",
            "title": "Overallcoveragerate"
          },
          "gapCount": {
            "type": "integer",
            "title": "Gapcount"
          },
          "partialCount": {
            "type": "integer",
            "title": "Partialcount"
          },
          "orphanNodeCount": {
            "type": "integer",
            "title": "Orphannodecount"
          },
          "diagnosticsMode": {
            "type": "string",
            "title": "Diagnosticsmode"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticFindingResponse"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "graphVersion",
          "generatedAt",
          "overallCoverageRate",
          "gapCount",
          "partialCount",
          "orphanNodeCount",
          "diagnosticsMode",
          "findings"
        ],
        "title": "GraphDiagnosticReportResponse"
      },
      "GraphEdgeReviewRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "GraphEdgeReviewRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecognitionCandidateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "course": {
            "type": "string",
            "title": "Course"
          },
          "candidateType": {
            "$ref": "#/components/schemas/RecognitionCandidateType"
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence"
          },
          "risk": {
            "$ref": "#/components/schemas/RecognitionCandidateRisk"
          },
          "sourceNode": {
            "type": "string",
            "title": "Sourcenode"
          },
          "relation": {
            "type": "string",
            "title": "Relation"
          },
          "targetNode": {
            "type": "string",
            "title": "Targetnode"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "processorVersion": {
            "type": "string",
            "title": "Processorversion"
          },
          "generatedAt": {
            "type": "string",
            "title": "Generatedat"
          },
          "reviewStatus": {
            "$ref": "#/components/schemas/CandidateReviewStatus"
          },
          "impact": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Impact"
          },
          "supportStrength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supportstrength"
          },
          "conflictMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflictmessage"
          },
          "reviewedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewedby"
          },
          "reviewedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewedat"
          },
          "reviewComment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewcomment"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/CandidateEvidenceResponse"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "course",
          "candidateType",
          "confidence",
          "risk",
          "sourceNode",
          "relation",
          "targetNode",
          "explanation",
          "processorVersion",
          "generatedAt",
          "reviewStatus",
          "impact",
          "evidence"
        ],
        "title": "RecognitionCandidateResponse"
      },
      "RecognitionCandidateRisk": {
        "type": "string",
        "enum": [
          "highImpact",
          "lowConfidence",
          "conflict",
          "normal"
        ],
        "title": "RecognitionCandidateRisk"
      },
      "RecognitionCandidateType": {
        "type": "string",
        "enum": [
          "关系候选",
          "映射候选",
          "节点候选"
        ],
        "title": "RecognitionCandidateType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}